Implement dynamic permission checks for signature and vehicle management views
- Added permission checks for updating signatures in edit.blade.php.
- Restricted access to create signature button in index.blade.php.
- Moved signature table rendering to a separate table.blade.php file.
- Integrated dynamic permission checks for editing, activating, deactivating, and deleting signatures in table.blade.php.
- Refactored supply orders index and table views to utilize dynamic permission checks.
- Created a separate table.blade.php for supply orders to streamline the index view.
- Added permission checks for vehicle management actions in create, edit, index, and show views.
- Ensured that action buttons in vehicle management are only visible to users with the appropriate permissions.

+++ resources/views/admin/customer/shipto/shiptotable.blade.php
@@ -59,20 +59,24 @@
                                         aria-expanded="false"><i class="fas fa-ellipsis-v"></i></a>
                                     <div class="dropdown-menu dropdown-menu-end">
                                         <ul>
-                                            <li>
-                                                <a class="dropdown-item"
-                                                    href="{{ route('admin.customer.updateShipTo', $customer->id) }}"><i
-                                                        class="ti ti-edit fs_18 me-2"></i>Update</a>
-                                            </li>
-                                            <li>
-                                                <form action="{{ route('admin.customer.destroyShipTo', $customer->id) }}"
-                                                    method="POST" class="d-inline">
-                                                    @csrf
-                                                    <button type="button" class="dropdown-item"
-                                                        onclick="deleteData(this,'Wait! Are you sure you want to remove this ship to customer?')"><i
-                                                            class="far fa-trash-alt me-2"></i>Delete</button>
-                                                </form>
-                                            </li>
+                                            @can('has-dynamic-permission', 'ship_to_customers_list.edit')
+                                                <li>
+                                                    <a class="dropdown-item"
+                                                        href="{{ route('admin.customer.updateShipTo', $customer->id) }}"><i
+                                                            class="ti ti-edit fs_18 me-2"></i>Update</a>
+                                                </li>
+                                            @endcan
+                                            @can('has-dynamic-permission', 'ship_to_customers_list.delete')
+                                                <li>
+                                                    <form action="{{ route('admin.customer.destroyShipTo', $customer->id) }}"
+                                                        method="POST" class="d-inline">
+                                                        @csrf
+                                                        <button type="button" class="dropdown-item"
+                                                            onclick="deleteData(this,'Wait! Are you sure you want to remove this ship to customer?')"><i
+                                                                class="far fa-trash-alt me-2"></i>Delete</button>
+                                                    </form>
+                                                </li>
+                                            @endcan
                                         </ul>
                                     </div>
                                 </div>
